backend/requirements.txt
selenium>=4.16.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytz>=2024.1
